crypto/cathode-ray-thruster/setup/pyproject.toml
[tool.poetry]
name = "rsa"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pycryptodome = "^3.14.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pwntools = "^4.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
